Change type of n_args in init method from short to int.

+++ docs/sample_code/MYINST/MYINST.cpp
@@ -57,7 +57,7 @@ MYINST :: ~MYINST()
    the error and exit. If you just want to warn the user and keep going,
    call warn() with a message.
 */
-int MYINST :: init(float p[], short n_args)
+int MYINST :: init(float p[], int n_args)
 {
    float outskip, inskip, dur;
